Cleaner HWPE test integration, added DMA-based tensors initialization. TODO: understand why the core reads Xs during matrix checks.
