+++ hwpe/redmule/inc/golden.h
